add skeleton for HTML options page
The content of the old XUL options page has been modified a bit already, but it is far from done. The concept here is to duplicate all code. You can still open the old preference page via the tools menu, for comparison.

* new modules folder, where modern ES6 modules will be stored (for now i18n.mjs and utils.mjs)
* resources for the new HTML options are stored in a new options folder
* a few checkboxes have been fully migrated, as an example how to use the LegacyPrefs API to load and save the existing preferences
* there are no automatism any more, you have to load and save prefs on your own (but it is not difficult, as shown with the checkboxes)
* defaults do not need to be handled, they are set on first run and the LegacyPrefs.getPref() function will either return the default value, or the user value

Feedback from users is that they do not need a tabbing mechanism in the options, but simply  large page with all the options, organised by headings. For the xpunge add-on I did implement a tabbing mechanism, if you prefer that (see version 5.0 at https://addons.thunderbird.net/addon/xpunge/versions/)

The biggest work will be to find replacements for the non-standard richlistbox and tree elements. The xpunge add-on may also be of help here.